Updated calculator app
Basic calculator. I manual created the interface, rather than using tables and buttons in HTML. It was a good way to practice CSS, and grid specifically. The only problem with this app right now is that the negative function doesn't work correctly when dividing or subtracting. The eval function doesn't recognize is next to a minus sign or divide sign. To be fixed soon. Need to customize this more as well.